Update bash.md
Added <br/> at the end of line 5, 7, 8, 9, and 11 to have the link description display in a row from top to bottom as opposed to how it was in its original display format, which side by side one link after the other in line, similar to word wrap. Having the link selection displayed in order would make it easier to see the separation.